Upgrade the root volume type to gp3
gp3 is 20% cheaper, and the baseline configuration offers better
performance than gp2 for volumes smaller than 2TB.  It also allows the
volume size and IOPS to be configured separately, whereas the two are
intertwined with gp2.

Also remove the delete_on_termination line, since
delete_on_termination = true is the default, so there is no need to
specify it.